隔离 i18n 代码
